chore: add default path to organize imports
Add default path to organize imports in tsconfig.app.json file to avoid
relative paths in imports.
