+++ src/features/Header/pages/HeaderPage.tsx
@@ -1,8 +1,8 @@
-import { HeaderAddress } from "../components/HeaderAddress";
-import { HeaderBrand } from "../components/HeaderBrand";
-import { HeaderCart } from "../components/HeaderCart";
-import { HeaderContainer } from "../styles/HeaderContainer";
-import { HeaderGroup } from "../styles/HeaderGroup";
+import { HeaderAddress } from "@features/Header/components/HeaderAddress";
+import { HeaderBrand } from "@features/Header/components/HeaderBrand";
+import { HeaderCart } from "@features/Header/components/HeaderCart";
+import { HeaderContainer } from "@features/Header/styles/HeaderContainer";
+import { HeaderGroup } from "@features/Header/styles/HeaderGroup";
 
 export default function HeaderPage() {
   return (
