+++ src/features/Home/page/HomePage.tsx
@@ -1,8 +1,8 @@
-import { HomeBannerTitle } from "../components/HomeBannerTitle";
-import { HomeBannerImage } from "../components/HomeBannerImage";
-import { HomeContainer } from "../styles/HomeContainer";
-import { HomeContextFirst } from "../styles/HomeContextFirst";
-import { HomeContextSecond } from "../styles/HomeContextSecond";
+import { HomeBannerImage } from "@features/Home/components/HomeBannerImage";
+import { HomeBannerTitle } from "@features/Home/components/HomeBannerTitle";
+import { HomeContainer } from "@features/Home/styles/HomeContainer";
+import { HomeContextFirst } from "@features/Home/styles/HomeContextFirst";
+import { HomeContextSecond } from "@features/Home/styles/HomeContextSecond";
 
 export function HomePage() {
   return (
