+++ src/features/Home/styles/TitleBanner.ts
@@ -1,5 +1,5 @@
+import { device } from '@globalStyles/Devices';
 import styled from "styled-components";
-import { device } from "../../../styles/Devices";
 
 export const TitleBanner = styled.h1`
   text-align: center;
