+++ src/layouts/DefaultLayout.tsx
@@ -1,6 +1,6 @@
+import { LayoutContainer } from "@globalStyles/LayoutContainer";
 import { Outlet } from "react-router-dom";
-import { LayoutContainer } from "../styles/LayoutContainer";
-import HeaderPage from "../features/Header/pages/HeaderPage";
+import HeaderPage from "@features/Header/pages/HeaderPage";
 
 export function DefaultLayout() {
   return (
