+++ src/types/ColorType.ts
@@ -5,10 +5,3 @@ export enum ColorType {
   INFO = "info",
   DEFAULT = "default"
 }
-
-export enum ColorVariant {
-  ORANGE = "orange",
-  YELLOW = "yellow",
-  BLACK = "black",
-  PURPLE = "purple"
-}
