+++ src/types/ColorVariant.ts
@@ -1,0 +1,6 @@
+export enum ColorVariant {
+  ORANGE = "orange",
+  YELLOW = "yellow",
+  BLACK = "black",
+  PURPLE = "purple"
+}
